extract PriorityQueue / Ordering from SortingTopNProjector
so that it can be re-used in other places
